# backend/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/team/create": {
            "post": {
                "description": "创建 team, 并且创建人加入了 team, 后端随机生成 teamName, 存入了数据库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Create a new team",
                "parameters": [
                    {
                        "description": "Create Team form",
                        "name": "createTeamForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.CreateTeamForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CreateTeamResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"User already belongs to a team, cannot create team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"User not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"Failed to create team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/get/list": {
            "get": {
                "description": "Get all teams",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get Team List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.TeamListResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Failed to fetch teams\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/get/preferences/{userId}": {
            "get": {
                "description": "Get the team preferences for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project Preference"
                ],
                "summary": "Get team preferences",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.PreferenceResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"User not found\"}\" or {\"error\": \"User does not belong to any team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/get/student-info/{teamName}": {
            "get": {
                "description": "Get student list information by team name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get student list information by team name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team Name",
                        "name": "teamName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.StudentInfoResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Team not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/invite/{userId}/{teamId}": {
            "get": {
                "description": "Invite a student to join a team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Invite User to Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"User invited to team successfully\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"User already belongs to a team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"User not found\"} or gin.H{\"error\": \"Team not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Failed to invite user to team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/join": {
            "put": {
                "description": "用户加入团队",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Join a team",
                "parameters": [
                    {
                        "description": "Join Team form",
                        "name": "joinTeamForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.JoinTeamForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.JoinTeamResponse"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"Validation failed\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"Team not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "{\"error\": \"User already belongs to a team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"Failed to update team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/team/leave/{userId}": {
            "delete": {
                "description": "用户离开团队",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Leave a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"User has left the team successfully\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"User does not belong to any team\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"User not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"Failed to update user\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/team/profile/{userId}": {
            "get": {
                "description": "获取用户所属团队的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get Team Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetTeamProfileResponse"
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"Team not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.CreateTeamForm": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "forms.JoinTeamForm": {
            "type": "object",
            "required": [
                "teamIdShow",
                "userId"
            ],
            "properties": {
                "teamIdShow": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "response.CreateTeamResponse": {
            "type": "object",
            "properties": {
                "teamId": {
                    "type": "integer"
                },
                "teamIdShow": {
                    "type": "integer"
                },
                "teamMember": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TeamMember"
                    }
                },
                "teamName": {
                    "type": "string"
                }
            }
        },
...
        "response.PreferenceResponse": {
            "type": "object",
            "properties": {
                "projectId": {
                    "type": "integer"
                },
                "projectTitle": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "response.StudentInfoResponse": {
            "type": "object",
            "properties": {
                "avatarURL": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.TeamListResponse": {
            "type": "object",
            "properties": {
                "teamId": {
                    "type": "integer"
                },
                "teamName": {
                    "type": "string"
                },
                "teamSkills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}